Added about us page & made changes to the layout.
The about us page, contains information about why the website was
created, and what it aims to achieve. I also changed the layout, making
the main section a white background, and adding a spiral background on
the sides. THE 'images/spirals.gif' FILE, WILL HAVE TO BE REMADE, AS IT IS NOT IN THE PUBLIC DOMAIN. IT IS ONLY THERE AS REFERENCE.