Fixed issues with logo
There were parts of the logo, were the stroke was missing, leaving white
space. I fixed it, and now the stroke is full all arround.
I also made the text into a vector object, so that it can be displayed
across all web browsers, that don't have the text's original font installed.